UI and Managers
Added a modular button for UI to be used, this has now implemented into all the UI stuff.
Filled with adding in a end game state that sort of work. (I've made comments so you can find them).
Tweaked some of the UI animations for the elements on the screen.
Added in some new sounds but nothing big